Add action section and footer
Created the action card, the content for it, the footer
and the content of the footer including a date that is
a javascript snippet to automatically populate the
current year. Styled the footer and action section with flexbox
and corrected some of how the buttons are displayed
to handle those html tags better in the page.